requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
accelerate=0.10.0=pypi_0
adapter-transformers=3.0.1=pypi_0
aiohttp=3.8.1=py37h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd3eb1b0_0
bashutils=0.0.4=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
boto3=1.24.79=pypi_0
botocore=1.27.79=pypi_0
bottleneck=1.3.5=py37h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.3.2=pypi_0
dbus=1.13.18=hb2f20db_0
dictdiffer=0.8.0=py_0
dill=0.3.4=pypi_0
evaluate=0.2.2=pypi_0
expat=2.4.4=h295c915_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.7.1=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py37h7f8727e_0
fsspec=2022.5.0=pypi_0
gdown=4.5.1=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.1=py_2
gputil=1.4.0=pypi_0
grpcio=1.42.0=py37hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.8.1=pypi_0
huggingface_hub=0.2.1=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=pypi_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jmespath=1.0.1=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonlines=2.0.0=pyhd3eb1b0_2
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
layers=0.1.5=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=5.2.0=py37h7f8727e_2
multiprocess=0.70.12.2=pypi_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.0=pyhd3eb1b0_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pcre=8.45=h295c915_0
pillow=9.2.0=py37hace64e9_1
pip=22.1.2=py37h06a4308_0
ply=3.11=py37_0
protobuf=3.20.1=py37h295c915_0
psutil=5.9.1=pypi_0
pyarrow=8.0.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pyjwt=2.4.0=py37h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37_1
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.6.4=pypi_0
pytorch-mutex=1.0=cuda
pytorch-transformers=1.2.0=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
quadprog=0.1.11=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.6.2=pypi_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.0=py_0
responses=0.18.0=pypi_0
rouge-score=0.0.4=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.6.0=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37hc147768_0
seaborn=0.12.2=py37h06a4308_0
sentencepiece=0.1.97=pypi_0
seqeval=1.2.2=pypi_0
setuptools=61.2.0=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.2=h5082296_0
tabulate=0.8.10=py37h06a4308_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py37hca6d32c_0
tensorboard-plugin-wit=1.8.1=py37h06a4308_0
tensorboardx=2.2=pyhd3eb1b0_0
termcolor=1.1.0=py37h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.11.0=py37_cu113
torchmetrics=0.9.1=pypi_0
torchvision=0.12.0=py37_cu113
tornado=6.2=py37h5eee18b_0
tqdm=4.64.0=py37h06a4308_0
transformers=4.17.0=pypi_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.11=py37h06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.8.0=py37h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
